refactored test in another file and wat files
