+++ circuit/src/execution-delivery/fixtures/code_call.wat
@@ -1,0 +1,43 @@
+(module
+;; seal_call(
+;;    callee_ptr: u32,
+;;    callee_len: u32,
+;;    gas: u64,
+;;    value_ptr: u32,
+;;    value_len: u32,
+;;    input_data_ptr: u32,
+;;    input_data_len: u32,
+;;    output_ptr: u32,
+;;    output_len_ptr: u32
+;;) -> u32
+(import "seal0" "seal_call" (func $seal_call (param i32 i32 i64 i32 i32 i32 i32 i32 i32) (result i32)))
+(import "env" "memory" (memory 1 1))
+(func (export "call")
+    (drop
+        (call $seal_call
+            (i32.const 4)  ;; Pointer to "callee" address.
+            (i32.const 32)  ;; Length of "callee" address.
+            (i64.const 0)  ;; How much gas to devote for the execution. 0 = all.
+            (i32.const 36) ;; Pointer to the buffer with value to transfer
+            (i32.const 8)  ;; Length of the buffer with value to transfer.
+            (i32.const 44) ;; Pointer to input data buffer address
+            (i32.const 4)  ;; Length of input data buffer
+            (i32.const 4294967295) ;; u32 max value is the sentinel value: do not copy output
+            (i32.const 0) ;; Length is ignored in this case
+        )
+    )
+)
+(func (export "deploy"))
+
+;; Destination AccountId (ALICE)
+(data (i32.const 4)
+    "\01\01\01\01\01\01\01\01\01\01\01\01\01\01\01\01"
+    "\01\01\01\01\01\01\01\01\01\01\01\01\01\01\01\01"
+)
+
+;; Amount of value to transfer.
+;; Represented by u64 (8 bytes long) in little endian.
+(data (i32.const 36) "\06\00\00\00\00\00\00\00")
+
+(data (i32.const 44) "\01\02\03\04")
+)
