+++ circuit/src/execution-delivery/fixtures/invalid_code.wat
@@ -1,0 +1,1 @@
+ invalid code
